test_requirements.txt
nose2
nose-cov
six
